feat: Market data from Alpaca API (#18)
* Working on stuff.

* feat: Doing stuff.

* refactor: Cleaning up.

* fix: Added `Distinct()` to interesting symbols return.

* test: Updated HTTP file.

---------

+++ TradingBot/Models/DailyPricePrediction.cs
@@ -3,7 +3,7 @@
 public sealed class DailyPricePrediction
 {
     public required DateOnly Date { get; init; }
-    public required double ClosingPrice { get; init; }
-    public required double HighPrice { get; init; }
-    public required double LowPrice { get; init; }
+    public required decimal ClosingPrice { get; init; }
+    public required decimal HighPrice { get; init; }
+    public required decimal LowPrice { get; init; }
 }
